+++ lib/src/main/java/com/github/thibaultbee/srtdroid/enums/KMState.kt
@@ -15,10 +15,35 @@
  */
 package com.github.thibaultbee.srtdroid.enums
 
+/**
+ * The defined encryption state as performed by the Key Material Exchange, used by
+ * [SockOpt.RCVKMSTATE], [SockOpt.SNDKMSTATE] and [SockOpt.KMSTATE] options
+ *
+ * **See Also:** [srt_km_state](https://github.com/Haivision/srt/blob/master/docs/APISocketOptions.md#srt_km_state)
+ */
 enum class KMState {
-    KM_S_UNSECURED,      //No encryption
-    KM_S_SECURING,       //Stream encrypted, exchanging Keying Material
-    KM_S_SECURED,        //Stream encrypted, keying Material exchanged, decrypting ok.
-    KM_S_NOSECRET,       //Stream encrypted and no secret to decrypt Keying Material
-    KM_S_BADSECRET       //Stream encrypted and wrong secret, cannot decrypt Keying Material
+    /**
+     * No encryption
+     */
+    KM_S_UNSECURED,
+
+    /**
+     * Stream encrypted, exchanging Keying Material
+     */
+    KM_S_SECURING,
+
+    /**
+     * Stream encrypted, keying Material exchanged, decrypting ok.
+     */
+    KM_S_SECURED,
+
+    /**
+     * Stream encrypted and no secret to decrypt Keying Material
+     */
+    KM_S_NOSECRET,
+
+    /**
+     * Stream encrypted and wrong secret, cannot decrypt Keying Material
+     */
+    KM_S_BADSECRET
 }